requirementes.txt
APScheduler==3.9.1.post1
Django~=3.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
dj-database-url==1.2.0
psycopg2-binary==2.9.5
whitenoise[brotli]==6.2.0
gunicorn==20.1.0
djangorestframework